Convert HelpButton to TypeScript
